# Optimize chunk comparisons in wildcard matching