--- a/src/redisearch_rs/wildcard/Cargo.toml
+++ b/src/redisearch_rs/wildcard/Cargo.toml
@@ -11,6 +11,9 @@ harness = false
 [lints]
 workspace = true
 
+[dependencies]
+memchr.workspace = true
+
 [dev-dependencies]
 criterion.workspace = true
 proptest = { workspace = true, features = ["std"] }