@LukeMathWalker 2025-05-26
> Looks good to me, could replicate the same relative benchmark gains locally too. Only question is the "* vs randomkey" regression 🤔 how common is that use case? I guess given the other gains this would be a worthwhile tradeoff but 11% is a relevant regression

That's a degenerate case—if you want all entries, you should do a plain search rather than a wildcard search. So I'd say it's an acceptable trade-off.